requirements.txt
pyserial
pyyaml